More work on automatic rotation
