+++ lemur/certificates/service.py
@@ -1,28 +1,33 @@
 """
-.. module: service
+.. module: lemur.certificate.service
     :platform: Unix
     :copyright: (c) 2015 by Netflix Inc., see AUTHORS for more
     :license: Apache, see LICENSE for more details.
 .. moduleauthor:: Kevin Glisson <kglisson@netflix.com>
 """
 import arrow
-from cryptography import x509
-from cryptography.hazmat.backends import default_backend
-from cryptography.hazmat.primitives import hashes, serialization
+from datetime import timedelta
+
 from flask import current_app
 from sqlalchemy import func, or_
 
+from cryptography import x509
+from cryptography.hazmat.backends import default_backend
+from cryptography.hazmat.primitives import hashes, serialization
+
 from lemur import database
-from lemur.authorities.models import Authority
-from lemur.certificates.models import Certificate
-from lemur.common.utils import generate_private_key
-from lemur.destinations.models import Destination
-from lemur.domains.models import Domain
 from lemur.extensions import metrics
-from lemur.notifications.models import Notification
 from lemur.plugins.base import plugins
-from lemur.roles import service as role_service
+from lemur.common.utils import generate_private_key
+
 from lemur.roles.models import Role
+from lemur.domains.models import Domain
+from lemur.authorities.models import Authority
+from lemur.destinations.models import Destination
+from lemur.certificates.models import Certificate
+from lemur.notifications.models import Notification
+
+from lemur.roles import service as role_service
 
 
 def get(cert_id):
@@ -71,6 +76,24 @@ def get_by_source(source_label):
     :return:
     """
     return Certificate.query.filter(Certificate.sources.any(label=source_label))
+
+
+def get_all_pending_rotation():
+    """
+    Retrieves all certificates that need to be rotated.
+
+    Must be X days from expiration, uses `LEMUR_DEFAULT_ROTATION_INTERVAL`
+    to determine how many days from expiration the certificate must be
+    for rotation to be pending.
+
+    :return:
+    """
+    now = arrow.utcnow()
+    interval = current_app.config.get('LEMUR_DEFAULT_ROTATION_INTERVAL', 30)
+    end = now + timedelta(days=interval)
+
+    return Certificate.query.filter(Certificate.rotation == True)\
+        .filter(Certificate.not_after <= end.format('YYYY-MM-DD')).all()  # noqa
 
 
 def find_duplicates(cert):
@@ -523,9 +546,9 @@ def reissue_certificate(certificate, replace=None, user=None):
     else:
         primitives['creator'] = user
 
-    new_cert = create(**primitives)
-
     if replace:
-        certificate.notify = False
+        primitives['replaces'] = certificate
+
+    new_cert = create(**primitives)
 
     return new_cert
